Worked to return the functionality to the game.

+++ Adventure/src/computc/entities/BigThug.java
@@ -8,7 +8,7 @@ import computc.worlds.rooms.Room;
 
 public class BigThug extends Thug
 {
-	public BigThug(Dungeon dungeon, Room room, float x, float y) throws SlickException
+	public BigThug(Dungeon dungeon, Room room, int x, int y) throws SlickException
 	{
 		super(dungeon, room, x, y);
