+++ Adventure/src/computc/entities/Enemy.java
@@ -7,7 +7,7 @@ import computc.worlds.rooms.Room;
 
 public abstract class Enemy extends Entity
 {
-	public Enemy(Dungeon dungeon, Room room, float x, float y)
+	public Enemy(Dungeon dungeon, Room room, int x, int y)
 	{
 		super(dungeon, room, x, y);
 	}
